Set a good starting point for distance only fitness
Basic logic is that at a maximum of 3 extrusion widths away (0.8 *3 = 24) per edge would be really bad. Constraining too much tighter is not a great choice.
But if all edges are greater than that the fitness should be 0. On average most edges in a good amount will be about or less than an extrusion width.

This math will need updated for future alignment math on the fitness