DemoFitness - lambda for fitness and weight
The weight and fitness functions of a DemoFitness are now described as lambda function in the controller which are passed on initialization.
Potential drawback - distance is not longer a static function for calculation since it requires the input, but that isn't too bad computationally.
Huge advantage -> easier for user to control and can be incorporated as a run parameter to document